backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.25
langgraph>=0.0.26
pydantic>=2.5.3
python-dotenv==1.0.0
httpx>=0.26.0
